Redesign sidenav bar and rework one IN standard card
Instead of linking the sidenav bar to each of the individual
standards, it now contains links the overarching concepts
(ie. Data and Information). Furthermore, this commit
changes the structure of the first standard card on the page
to follow the following structure:

<div class='accordion'>
    <h2>Standard Title</h2>
    <p>Short description</p>
    <div class = "accordion-item">
        <h3>Title<h3>
        <div class="accordion-content">
            <*
            <h2></h2>
            /*>
        </div>
    </div>
</div>

Future commits will change the entire page to this format